+++ changelog.md
@@ -1,4 +1,8 @@
-### 2.5.8 (2023-10-08)
+### 2.5.9 (2023-11-01)
+- added bGates module
+- bTogglerCompact: added "Trigger to Gate Output" option in context menu
+
+### 2.5.8 (2023-10-07)
 - added sickoAmp module
 - clocker: added swing feature
 - sickoPlayer / sickoSampler / sickoSampler2: added EoC pulse when sample has reached begin/end if start/end cursors are set to 0% or 100%